resolvendo alguns problemas criando logger e apresentacao

+++ codigoFinal/src/com/academia/sistema/service/AlunoService.java
@@ -1,16 +1,17 @@
 package com.academia.sistema.service;
 
 import java.util.List;
+import com.academia.sistema.model.Aluno;
+import com.academia.sistema.dao.DAO;
+import com.academia.sistema.dao.ConectorDAO;
+import com.academia.sistema.util.CPFValidatorController;
 
-import com.academia.sistema.model.*;
-import com.academia.sistema.dao.*;
-import com.academia.sistema.util.*;
-
-public class AlunoService { 
+public class AlunoService {
     private DAO<Aluno> alunoDAO = ConectorDAO.getInstancia().criarAlunoDAO();
+    private CPFValidatorController cpfValidator = CPFValidatorController.getInstancia();
 
     public void adicionarAluno(Aluno aluno) {
-        if (CPFValidatorUtil.validarCPF(aluno.getCpf())) {
+        if (cpfValidator.validarCPF(aluno.getCpf())) {
             alunoDAO.adicionar(aluno);
         } else {
             throw new IllegalArgumentException("CPF inválido");
@@ -18,11 +19,15 @@ public class AlunoService {
     }
 
     public Aluno buscarAluno(String cpf) {
-        return alunoDAO.buscar(cpf);
+        if (cpfValidator.validarCPF(cpf)) {
+            return alunoDAO.buscar(cpf);
+        } else {
+            throw new IllegalArgumentException("CPF inválido");
+        }
     }
 
     public void atualizarAluno(Aluno aluno) {
-        if (CPFValidatorUtil.validarCPF(aluno.getCpf())) {
+        if (cpfValidator.validarCPF(aluno.getCpf())) {
             alunoDAO.atualizar(aluno);
         } else {
             throw new IllegalArgumentException("CPF inválido");
@@ -30,7 +35,11 @@ public class AlunoService {
     }
 
     public void removerAluno(String cpf) {
-        alunoDAO.remover(cpf);
+        if (cpfValidator.validarCPF(cpf)) {
+            alunoDAO.remover(cpf);
+        } else {
+            throw new IllegalArgumentException("CPF inválido");
+        }
     }
 
     public List<Aluno> listarAlunos() {
